backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# RAG相关
langchain==0.0.350
langchain-community==0.0.10
chromadb==0.4.18
sentence-transformers==2.2.2
transformers==4.36.0
torch>=2.6.0

# 图像处理
opencv-python==4.8.1.78
Pillow>=10.1.0
numpy>=1.24.3
paddlepaddle==2.5.2
paddleocr==2.7.0.3

# 数据处理
pandas==2.1.4
scipy==1.11.4
scikit-learn==1.3.2

# 文档处理
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1

# 数据库
sqlalchemy==2.0.23
alembic==1.13.1

# 工具库
requests==2.31.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# 日志
loguru==0.7.2

# 其他
httpx==0.25.2
jinja2==3.1.2